+++ example/src/main/kotlin/com/example/afterpay/data/Product.kt
@@ -1,10 +1,11 @@
 package com.example.afterpay.data
 
+import java.math.BigDecimal
 import java.util.UUID
 
 data class Product(
     val id: UUID,
     val name: String,
     val description: String,
-    val price: Double
+    val price: BigDecimal
 )
